schema generation: restructure files and do a tiny bit of refactoring
...with a view to being able to present a vaguely similar interface to
the assessment schema generator
